+++ src/main/kotlin/nextstep/mission/LottoApplication.kt
@@ -4,7 +4,7 @@ import nextstep.mission.lotto.Lotto
 import nextstep.mission.lotto.LottoShop
 import nextstep.mission.lotto.io.ConsoleInput
 import nextstep.mission.lotto.io.ConsoleOutput
-import nextstep.mission.lotto.io.dto.LottoDto
+import nextstep.mission.lotto.dto.LottoDto
 import nextstep.mission.lotto.vo.LottoNumbers
 import nextstep.mission.lotto.vo.WinningResult
